Enum relay_common::EventType

pub enum EventType {
    Error,
    Csp,
    Hpkp,
    ExpectCt,
    ExpectStaple,
    Transaction,
    Default,
}

The type of an event.

The event type determines how Sentry handles the event and has an impact on processing, rate limiting, and quotas. There are three fundamental classes of event types:

Error monitoring events (default, error): Processed and grouped into unique issues based on their exception stack traces and error messages.

Security events (csp, hpkp, expectct, expectstaple): Derived from Browser security violation reports and grouped into unique issues based on the endpoint and violation. SDKs do not send such events.

Transaction events (transaction): Contain operation spans and collected into traces for performance monitoring.

Variants

Error

Events that carry an exception payload.

Csp

A CSP violation payload.

Hpkp

An HPKP violation payload.

ExpectCt

An ExpectCT violation payload.

ExpectStaple

An ExpectStaple violation payload.

Transaction

Performance monitoring transactions carrying spans.

Default

All events that do not qualify as any other type.