Enum relay_common::SpanStatus

#[repr(u8)]pub enum SpanStatus {
    Ok,
    Cancelled,
    Unknown,
    InvalidArgument,
    DeadlineExceeded,
    NotFound,
    AlreadyExists,
    PermissionDenied,
    ResourceExhausted,
    FailedPrecondition,
    Aborted,
    OutOfRange,
    Unimplemented,
    InternalError,
    Unavailable,
    DataLoss,
    Unauthenticated,
}

Trace status.

Values from https://github.com/open-telemetry/opentelemetry-specification/blob/8fb6c14e4709e75a9aaa64b0dbbdf02a6067682a/specification/api-tracing.md#status Mapping to HTTP from https://github.com/open-telemetry/opentelemetry-specification/blob/8fb6c14e4709e75a9aaa64b0dbbdf02a6067682a/specification/data-http.md#status

Variants

Ok

The operation completed successfully.

HTTP status 100..299 + successful redirects from the 3xx range.

Cancelled

The operation was cancelled (typically by the user).

Unknown

Unknown. Any non-standard HTTP status code.

“We do not know whether the transaction failed or succeeded”

InvalidArgument

Client specified an invalid argument. 4xx.

Note that this differs from FailedPrecondition. InvalidArgument indicates arguments that are problematic regardless of the state of the system.

DeadlineExceeded

Deadline expired before operation could complete.

For operations that change the state of the system, this error may be returned even if the operation has been completed successfully.

HTTP redirect loops and 504 Gateway Timeout

NotFound

404 Not Found. Some requested entity (file or directory) was not found.

AlreadyExists

Already exists (409)

Some entity that we attempted to create already exists.

PermissionDenied

403 Forbidden

The caller does not have permission to execute the specified operation.

ResourceExhausted

429 Too Many Requests

Some resource has been exhausted, perhaps a per-user quota or perhaps the entire file system is out of space.

FailedPrecondition

Operation was rejected because the system is not in a state required for the operation’s execution

Aborted

The operation was aborted, typically due to a concurrency issue.

OutOfRange

Operation was attempted past the valid range.

Unimplemented

501 Not Implemented

Operation is not implemented or not enabled.

InternalError

Other/generic 5xx.

Unavailable

503 Service Unavailable

DataLoss

Unrecoverable data loss or corruption

Unauthenticated

401 Unauthorized (actually does mean unauthenticated according to RFC 7235)

Prefer PermissionDenied if a user is logged in.