relay_common
Module glob2

Serializable glob patterns for the API.

Structs

Glob

A simple glob matcher.

GlobBuilder

GlobBuilder provides the posibility to fine tune the final Glob, mainly what capture groups will be enabled in the underlying regex.

GlobMatcher

Helper for glob matching

LazyGlob

Wrapper type around the raw string pattern and the Glob.