relay_common

Macro impl_str_ser

macro_rules! impl_str_ser {
    ($type:ty) => { ... };
}

Helper macro to implement string based serialization.

If a type implements Display then this automatically implements a serializer for that type that dispatches appropriately.