relay_common
Macro impl_str_serde

macro_rules! impl_str_serde {
    ($type:ty, $expectation:expr) => { ... };
}

Helper macro to implement string based serialization and deserialization.

If a type implements FromStr and Display then this automatically implements a serializer/deserializer for that type that dispatches appropriately.