Trait relay_common::metrics::prelude::Counted

pub trait Counted {
    fn count_with_tags(
        &'a self, 
        key: &'a str, 
        count: i64
    ) -> MetricBuilder<'a, 'a, Counter>;

    fn incr(&self, key: &str) -> Result<Counter, MetricError> { ... }

    fn incr_with_tags(&'a self, key: &'a str) -> MetricBuilder<'a, 'a, Counter> { ... }

    fn decr(&self, key: &str) -> Result<Counter, MetricError> { ... }

    fn decr_with_tags(&'a self, key: &'a str) -> MetricBuilder<'a, 'a, Counter> { ... }

    fn count(&self, key: &str, count: i64) -> Result<Counter, MetricError> { ... }
}

Trait for incrementing and decrementing counters.

Counters are simple values incremented or decremented by a client. The rates at which these events occur or average values will be determined by the server receiving them. Examples of counter uses include number of logins to a system or requests received.

See the Statsd spec for more information.

Note that tags are a Datadog extension to Statsd and may not be supported by your server.

Required methods

fn count_with_tags(
&'a self,
key: &'a str,
count: i64
) -> MetricBuilder<'a, 'a, Counter>

Increment or decrement the counter by the given amount and return a MetricBuilder that can be used to add tags to the metric.

Loading content...
Provided methods

fn incr(&self, key: &str) -> Result<Counter, MetricError>

Increment the counter by 1

fn incr_with_tags(&'a self, key: &'a str) -> MetricBuilder<'a, 'a, Counter>

Increment the counter by 1 and return a MetricBuilder that can be used to add tags to the metric.

fn decr(&self, key: &str) -> Result<Counter, MetricError>

Decrement the counter by 1

fn decr_with_tags(&'a self, key: &'a str) -> MetricBuilder<'a, 'a, Counter>

Decrement the counter by 1 and return a `MetricBuilder that can be used to add tags to the metric.

fn count(&self, key: &str, count: i64) -> Result<Counter, MetricError>

Increment or decrement the counter by the given amount

Loading content...