Trait relay_common::metrics::prelude::Gauged

pub trait Gauged {
    pub fn gauge_with_tags(
        &'a self, 
        key: &'a str, 
        value: u64
    ) -> MetricBuilder<'a, 'a, Gauge>;

    pub fn gauge(&self, key: &str, value: u64) -> Result<Gauge, MetricError> { ... }
}

Trait for recording gauge values.

Gauge values are an instantaneous measurement of a value determined by the client. They do not change unless changed by the client. Examples include things like load average or how many connections are active.

See the Statsd spec for more information.

Note that tags are a Datadog extension to Statsd and may not be supported by your server.

Required methods

pub fn gauge_with_tags(
&'a self,
key: &'a str,
value: u64
) -> MetricBuilder<'a, 'a, Gauge>

Record a gauge value with the given key and return a MetricBuilder that can be used to add tags to the metric.

Loading content...
Provided methods

pub fn gauge(&self, key: &str, value: u64) -> Result<Gauge, MetricError>

Record a gauge value with the given key

Loading content...