Trait relay_common::metrics::prelude::Setted

pub trait Setted {
    fn set_with_tags(
        &'a self, 
        key: &'a str, 
        value: i64
    ) -> MetricBuilder<'a, 'a, Set>;

    fn set(&self, key: &str, value: i64) -> Result<Set, MetricError> { ... }
}

Trait for recording set values.

Sets count the number of unique elements in a group. You can use them to, for example, count the unique visitors to your site.

See the Statsd spec for more information.

Required methods

fn set_with_tags(
&'a self,
key: &'a str,
value: i64
) -> MetricBuilder<'a, 'a, Set>

Record a single set value with the given key and return a MetricBuilder that can be used to add tags to the metric.

Loading content...
Provided methods

fn set(&self, key: &str, value: i64) -> Result<Set, MetricError>

Record a single set value with the given key

Loading content...