Trait relay_common::metrics::prelude::Timed

pub trait Timed {
    pub fn time_with_tags(
        &'a self, 
        key: &'a str, 
        time: u64
    ) -> MetricBuilder<'a, 'a, Timer>;

    pub fn time_duration_with_tags(
        &'a self, 
        key: &'a str, 
        duration: Duration
    ) -> MetricBuilder<'a, 'a, Timer>;

    pub fn time(&self, key: &str, time: u64) -> Result<Timer, MetricError> { ... }

    pub fn time_duration(
        &self, 
        key: &str, 
        duration: Duration
    ) -> Result<Timer, MetricError> { ... }
}

Trait for recording timings in milliseconds.

Timings are a positive number of milliseconds between a start and end time. Examples include time taken to render a web page or time taken for a database call to return.

See the Statsd spec for more information.

Note that tags are a Datadog extension to Statsd and may not be supported by your server.

Required methods

pub fn time_with_tags(
&'a self,
key: &'a str,
time: u64
) -> MetricBuilder<'a, 'a, Timer>

Record a timing in milliseconds with the given key and return a MetricBuilder that can be used to add tags to the metric.

pub fn time_duration_with_tags(
&'a self,
key: &'a str,
duration: Duration
) -> MetricBuilder<'a, 'a, Timer>

Record a timing in milliseconds with the given key and return a MetricBuilder that can be used to add tags to the metric.

The duration will be truncated to millisecond precision. If the duration cannot be represented as a u64 an error will be deferred and returned when MetricBuilder::try_send() is called.

Loading content...
Provided methods

pub fn time(&self, key: &str, time: u64) -> Result<Timer, MetricError>

Record a timing in milliseconds with the given key

pub fn time_duration(
&self,
key: &str,
duration: Duration
) -> Result<Timer, MetricError>

Record a timing in milliseconds with the given key

The duration will be truncated to millisecond precision. If the duration cannot be represented as a u64 an error will be returned.

Loading content...