Crate relay_common

Common functionality for the sentry relay.

Macros

clone

A cloning alternative to a move closure.

impl_str_de

Helper macro to implement string based deserialization.

impl_str_ser

Helper macro to implement string based serialization.

impl_str_serde

Helper macro to implement string based serialization and deserialization.

Structs

Auth

Represents an auth header.

CustomUnit

Custom user-defined units without builtin conversion.

Dsn

Represents a Sentry dsn.

GlobOptions

Controls the options of the globber.

ParseEventTypeError

An error used when parsing EventType.

ParseMetricUnitError

An error parsing a MetricUnit or one of its variants.

ParseProjectKeyError

An error parsing ProjectKey.

ParseSpanStatusError

Error parsing a SpanStatus.

ParseUnixTimestampError

An error returned from parsing UnixTimestamp.

ProjectId

The unique identifier of a Sentry project.

ProjectKey

The public key used in a DSN to identify and authenticate for a project at Sentry.

UnixTimestamp

A unix timestamp (full seconds elapsed since 1970-01-01 00:00 UTC).

Uuid

A Universally Unique Identifier (UUID).

Enums

DataCategory

Classifies the type of data that is being ingested.

DurationUnit

Time duration units used in MetricUnit::Duration.

EventType

The type of an event.

FractionUnit

Units of fraction used in MetricUnit::Fraction.

InformationUnit

Size of information derived from bytes, used in MetricUnit::Information.

MetricUnit

The unit of measurement of a metric value.

MonotonicResult

The conversion result of UnixTimestamp::to_instant.

ParseAuthError

Represents an auth header parsing error.

ParseDsnError

Represents a dsn url parsing error.

ParseProjectIdError

Raised if a project ID cannot be parsed from a string.

Scheme

Represents the scheme of an url http/https.

SpanStatus

Trace status.

Constants

PROTOCOL_VERSION

The latest version of the protocol.

Functions

chrono_to_positive_millis

Returns the positive number of milliseconds contained by this Duration as f64.

codeowners_match_bytes

Returns true if the codeowners regex matches, false otherwise.

duration_to_millis

Returns the number of milliseconds contained by this Duration as f64.

glob_match

Performs a glob operation.

glob_match_bytes

Performs a glob operation on bytes.

instant_to_date_time

Converts an Instant into a DateTime.

instant_to_system_time

Converts an Instant into a SystemTime.

is_valid_metric_name

Validates a metric name. This is the statsd name, i.e. without type or unit.