Crate relay_common

Common functionality for the sentry relay.

Modules

glob

Binary glob pattern matching for the C-ABI.

glob2

Serializable glob patterns for the API.

glob3

Alternative implementation of serializable glob patterns.

time

Utilities to deal with date-time types. (DateTime, Instant, SystemTime, etc)

Macros

derive_fromstr_and_display

Implements FromStr and Display on a flat/C-like enum such that strings roundtrip correctly and all variants can be FromStr’d.

impl_str_de

Helper macro to implement string based deserialization.

impl_str_ser

Helper macro to implement string based serialization.

impl_str_serde

Helper macro to implement string based serialization and deserialization.

Structs

Auth

Represents an auth header.

Dsn

Represents a Sentry dsn.

Enums

ParseAuthError

Represents an auth header parsing error.

ParseDsnError

Represents a dsn url parsing error.

Scheme

Represents the scheme of an url http/https.