Macro relay_common::clone

macro_rules! clone {
    ($($n:ident ,)+ || $body:expr) => { ... };
    ($($n:ident ,)+ |$($p:pat),+| $body:expr) => { ... };
}

A cloning alternative to a move closure.

When one needs to use a closure with move semantics one often needs to clone and move some of the free variables. This macro automates the process of cloning and moving variables.

The following code:

let shared = Arc::new(Mutex::new(0));

let cloned = shared.clone();
std::thread::spawn(move || {
    *cloned.lock().unwrap() = 42
}).join();

assert_eq!(*shared.lock().unwrap(), 42);

Can be rewritten in a cleaner way by using the clone! macro like so:

use relay_common::clone;

let shared = Arc::new(Mutex::new(0));
std::thread::spawn(clone!(shared, || {
    *shared.lock().unwrap() = 42
})).join();

assert_eq!(*shared.lock().unwrap(), 42);