Trait relay_common::metrics::prelude::Counted

pub trait Counted<T> where
    T: ToCounterValue, {
    fn count_with_tags(
        &'a self, 
        key: &'a str, 
        count: T
    ) -> MetricBuilder<'a, 'a, Counter>;

    fn count(&self, key: &str, count: T) -> Result<Counter, MetricError> { ... }
}

Trait for incrementing and decrementing counters.

Counters are simple values incremented or decremented by a client. The rates at which these events occur or average values will be determined by the server receiving them. Examples of counter uses include number of logins to a system or requests received.

The following types are valid for counters:

i64

See the Statsd spec for more information.

Note that tags are a Datadog extension to Statsd and may not be supported by your server.

Required methods

fn count_with_tags(
&'a self,
key: &'a str,
count: T
) -> MetricBuilder<'a, 'a, Counter>

Increment or decrement the counter by the given amount and return a MetricBuilder that can be used to add tags to the metric.

Provided methods

fn count(&self, key: &str, count: T) -> Result<Counter, MetricError>

Increment or decrement the counter by the given amount