Module relay_common::time

Utilities to deal with date-time types. (DateTime, Instant, SystemTime, etc)

Structs

ParseUnixTimestampError

An error returned from parsing UnixTimestamp.

UnixTimestamp

A unix timestamp (full seconds elapsed since 1970-01-01 00:00 UTC).

Enums

MonotonicResult

The conversion result of UnixTimestamp::to_instant.

Functions

chrono_to_positive_millis

Returns the positive number of milliseconds contained by this Duration as f64.

duration_to_millis

Returns the number of milliseconds contained by this Duration as f64.

instant_to_date_time

Converts an Instant into a DateTime.

instant_to_system_time

Converts an Instant into a SystemTime.